Fix boot failure on CentOS without gateway
Seen on CentOS 8, but probably more related to cloud-init. For a host
with network configuration where at least one interface does not have a
gateway.

    is_ipv6 = subnet.get('ipv6') or is_ipv6_addr(route['gateway'])
    KeyError: 'gateway'

This happens because there is an entry in the routes list in
network_info.json that looks like this:

    {"network": "0.0.0.0", "netmask": "0.0.0.0"}

This change fixes the issue by only adding a default route if a gateway
has been specified for an interface.

Fixes: #6