Support VLAN links network_info.json
With older versions of cloud-init (e.g. CentOS 7), it seemed to be
possible to specify an interface named <link>.<vlan>, and cloud-init
would assume that this is a VLAN interface on <link> with VLAN ID
<vlan>.

In more recent versions of cloud-init (e.g. CentOS 8), several
additional parameters are required in more recent cloud-init versions:

* vlan_id
* vlan_link
* vlan_mac_address

While the MAC address is required, it does seem possible to set it to
null. Additionally, the backend device must be listed in the
configuration.